+++ components/shys_m5_dial/ha_device_mode_fan_speed.h
@@ -23,7 +23,7 @@ namespace esphome
                     
                 }
 
-                void sendValueToHomeAssistant(int value) override {
+                void sendValueToHomeAssistant(float value) override {
                     haApi.setFanSpeed(this->device.getEntityId(), value);
                 }
 
@@ -63,7 +63,7 @@ namespace esphome
                     this->stateIsOn = (strcmp(newState.c_str(), "on") == 0);
                 }
 
-                int getValue() override {
+                float getValue() override {
                     return this->stateIsOn ? this->value : 0;
                 }
                 
@@ -118,12 +118,12 @@ namespace esphome
                         if(this->isValueModified()){
                             return;
                         }
-                        auto val = parse_number<int>(state);
+                        auto val = parse_number<float>(state);
                         if (!val.has_value()) {
                             this->setReceivedValue(0);
                             ESP_LOGD("HA_API", "No Percentage value in %s for %s", state.c_str(), this->device.getEntityId().c_str());
                         } else {
-                            this->setReceivedValue(round((int)val.value()));
+                            this->setReceivedValue(round(val.value()));
                             ESP_LOGI("HA_API", "Got Percentage value %i for %s", val.value(), this->device.getEntityId().c_str());
                         }
                     });
@@ -144,7 +144,7 @@ namespace esphome
                     }
                 }
 
-                bool onTouch(M5DialDisplay& display, uint16_t x, uint16_t y) override {
+                bool onTouch(M5DialDisplay& display, float x, float y) override {
                     return this->defaultOnTouch(display, x, y);
                 }
